edited files for .gitignore and more

+++ products/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 class Product(models.Model):
     name = models.CharField(max_length=255)
-    price = models.FloatField()
+    price = models.DecimalField(decimal_places=2, max_digits=4)
     stock = models.IntegerField()
     image_url = models.CharField(max_length=2083)
